week-5/src/requirements.txt
# using v2 (not the latest v3) as this is good for windows, ubuntu and mac
psycopg2-binary
python-dotenv==0.15.0
tabulate
